+++ src/main/java/com/examplenewstack/newstack/collection/Collection.java
@@ -18,11 +18,10 @@ import java.util.List;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-
 public class Collection {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private int total_quantity;
@@ -36,6 +35,4 @@ public class Collection {
     // Relacionamento com book
     @OneToMany(mappedBy = "collection", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<Book> books;
-
-
 }
